.github/workflows: add repoman and pkgcheck
Currently, this overlay does not have any QA checks for continuous
integration. This means that updates can be merged that violate
Gentoo's QA, and existing violations are not publically seen and
so generally probably will not be dealt with.
Without this commit, this overlay will continue to lack QA checks.
With this commit, this overlay will have QA checks every time a push or
pull request is made on master. In addition, a QA check will occur
everyday. If you think a QA check occuring everyday is too much, I will
push another commit onto this branch that removes the daily checks.
This QA system is lifted from Gentoo's science overlay. I did not
include duplicates because there are no duplicates of these packages in
Gentoo's main overlay; however, I can add it if it is desired.
Finally, this overlay does not fail any QA checks, but it simply has
minor issues. Therefore, github should not show the overlay as currently
failing the QA check.
The same QA system in science:
https://github.com/gentoo/sci/tree/master/.github/workflows
https://github.com/gentoo/sci/tree/master/scripts
This commit was written and submitted by Lucas Mitrak.

Signed-off-by: Lucas Mitrak <lucas@lucasmitrak.com>